# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Service Assurance AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sites:
    get:
      operationId: getSites
      tags: [sites]
      summary: List sites
      parameters:
        - name: customerId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of sites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SiteWithCustomer"
    post:
      operationId: createSite
      tags: [sites]
      summary: Create a site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSiteRequest"
      responses:
        "201":
          description: Created site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
  /sites/{id}:
    get:
      operationId: getSite
      tags: [sites]
      summary: Get site by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteWithRelations"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateSite
      tags: [sites]
      summary: Update a site
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSiteRequest"
      responses:
        "200":
          description: Updated site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
    delete:
      operationId: deleteSite
      tags: [sites]
      summary: Delete a site
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /services:
    get:
      operationId: getServices
      tags: [services]
      summary: List services
      parameters:
        - name: customerId
          in: query
          schema:
            type: string
        - name: siteId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: vendorName
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceWithRelations"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get service by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceWithRelations"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceRequest"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /sla-policies:
    get:
      operationId: getSlaPolicies
      tags: [sla-policies]
      summary: List SLA policies
      responses:
        "200":
          description: SLA policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SlaPolicy"
    post:
      operationId: createSlaPolicy
      tags: [sla-policies]
      summary: Create a SLA policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSlaPolicyRequest"
      responses:
        "201":
          description: Created SLA policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlaPolicy"
  /sla-policies/{id}:
    put:
      operationId: updateSlaPolicy
      tags: [sla-policies]
      summary: Update a SLA policy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSlaPolicyRequest"
      responses:
        "200":
          description: Updated SLA policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlaPolicy"
    delete:
      operationId: deleteSlaPolicy
      tags: [sla-policies]
      summary: Delete a SLA policy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Customer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        accountNumber:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        primaryContactName:
          type: string
          nullable: true
        primaryContactEmail:
          type: string
          nullable: true
        primaryContactPhone:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - createdAt
        - updatedAt
...
    Site:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        siteName:
          type: string
        address1:
          type: string
          nullable: true
        address2:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        siteCode:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - siteName
        - createdAt
        - updatedAt
    SiteWithCustomer:
      allOf:
        - $ref: "#/components/schemas/Site"
        - type: object
          properties:
            customer:
              $ref: "#/components/schemas/Customer"
    SiteWithRelations:
      allOf:
        - $ref: "#/components/schemas/Site"
        - type: object
          properties:
            customer:
              $ref: "#/components/schemas/Customer"
            services:
              type: array
              items:
                $ref: "#/components/schemas/Service"
            tickets:
              type: array
              items:
                $ref: "#/components/schemas/Ticket"
    CreateSiteRequest:
      type: object
      properties:
        customerId:
          type: string
        siteName:
          type: string
        address1:
          type: string
          nullable: true
        address2:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        siteCode:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - customerId
        - siteName
    UpdateSiteRequest:
      type: object
      properties:
        siteName:
          type: string
        address1:
          type: string
          nullable: true
        address2:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        siteCode:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    Service:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        siteId:
          type: string
        vendorName:
          type: string
        serviceType:
          type: string
          enum: [DIA, Broadband, SD-WAN, Voice, Wireless, Other]
        circuitId:
          type: string
          nullable: true
        bandwidth:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, pending, down, impaired, disconnected]
        installDate:
          type: string
          nullable: true
        monthlyRecurringCharge:
          type: number
          nullable: true
        supportReference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - siteId
        - vendorName
        - serviceType
        - status
        - createdAt
        - updatedAt
    ServiceWithRelations:
      allOf:
        - $ref: "#/components/schemas/Service"
        - type: object
          properties:
            customer:
              $ref: "#/components/schemas/Customer"
            site:
              $ref: "#/components/schemas/Site"
    CreateServiceRequest:
      type: object
      properties:
        customerId:
          type: string
        siteId:
          type: string
        vendorName:
          type: string
        serviceType:
          type: string
          enum: [DIA, Broadband, SD-WAN, Voice, Wireless, Other]
        circuitId:
          type: string
          nullable: true
        bandwidth:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, pending, down, impaired, disconnected]
        installDate:
          type: string
          nullable: true
        monthlyRecurringCharge:
          type: number
          nullable: true
        supportReference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - customerId
        - siteId
        - vendorName
        - serviceType
        - status
    UpdateServiceRequest:
      type: object
      properties:
        vendorName:
          type: string
        serviceType:
          type: string
          enum: [DIA, Broadband, SD-WAN, Voice, Wireless, Other]
        circuitId:
          type: string
          nullable: true
        bandwidth:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, pending, down, impaired, disconnected]
        installDate:
          type: string
          nullable: true
        monthlyRecurringCharge:
          type: number
          nullable: true
        supportReference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    Ticket:
      type: object
      properties:
        id:
          type: string
        ticketNumber:
          type: string
        customerId:
          type: string
        siteId:
          type: string
          nullable: true
        serviceId:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        source:
          type: string
          enum: [manual, email, api]
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [new, investigating, vendor_engaged, dispatch_scheduled, monitoring, resolved, closed]
        outageType:
          type: string
          enum: [outage, impairment, informational, unknown]
        vendorTicketId:
          type: string
          nullable: true
        assignedToUserId:
          type: string
          nullable: true
        openedAt:
          type: string
          format: date-time
        lastUpdatedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        nextEscalationAt:
          type: string
          format: date-time
          nullable: true
        slaTargetMinutes:
          type: number
          nullable: true
        aiSummary:
          type: string
          nullable: true
        aiNormalizedStatus:
          type: string
          nullable: true
        aiCustomerUpdate:
          type: string
          nullable: true
        aiLastGeneratedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - ticketNumber
        - customerId
        - title
        - source
        - severity
        - status
        - outageType
        - openedAt
        - lastUpdatedAt
        - createdAt
        - updatedAt
...
    SlaPolicy:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        initialResponseMinutes:
          type: number
        escalationMinutes:
          type: number
        resolutionTargetMinutes:
          type: number
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - severity
        - initialResponseMinutes
        - escalationMinutes
        - resolutionTargetMinutes
        - isDefault
        - createdAt
        - updatedAt
    CreateSlaPolicyRequest:
      type: object
      properties:
        name:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        initialResponseMinutes:
          type: number
        escalationMinutes:
          type: number
        resolutionTargetMinutes:
          type: number
        isDefault:
          type: boolean
      required:
        - name
        - severity
        - initialResponseMinutes
        - escalationMinutes
        - resolutionTargetMinutes
        - isDefault
    UpdateSlaPolicyRequest:
      type: object
      properties:
        name:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        initialResponseMinutes:
          type: number
        escalationMinutes:
          type: number
        resolutionTargetMinutes:
          type: number
        isDefault:
          type: boolean
...