# lib/api-zod/src/generated/types/siteWithRelations.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Service Assurance AI API
 * OpenAPI spec version: 0.1.0
 */
import type { Customer } from "./customer";
import type { Service } from "./service";
import type { Site } from "./site";
import type { Ticket } from "./ticket";

export type SiteWithRelations = Site & {
  customer?: Customer;
  services?: Service[];
  tickets?: Ticket[];
};